requirements.txt
appdirs==1.4.4
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.3
distlib==0.3.1
fastapi==0.60.2
filelock==3.0.12
h11==0.9.0
httptools==0.1.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
joblib==0.16.0
more-itertools==8.4.0
numpy==1.19.1
packaging==20.4
pandas==1.1.0
pluggy==0.13.1
py==1.9.0
pydantic==1.6.1
pyparsing==2.4.7
pytest==6.0.1
pytest-cov==2.10.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
starlette==0.13.6
threadpoolctl==2.1.0
toml==0.10.1
tox==3.19.0
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
vertica-python==0.11.0
virtualenv==20.0.30
websockets==8.1
zipp==3.1.0
